+++ application/views/search.php
@@ -1,0 +1,8 @@
+<h1>Displaying Flights from {from} to {to}</h1>
+{bookings}
+    <h2>Flight path</h2>
+    {flights}
+        From {departureCommunity} to {destinationCommunity}</br>
+    {/flights}
+    <hr />
+{/bookings}